backend/requirements.txt
fastapi
uvicorn
pandas
requests
beautifulsoup4
aiosqlite
python-dotenv
pydantic
python-multipart
httpx
aiofiles
python-dateutil
numpy
